src/main/java/com/rkostiuk/cstask/service/DataJpaUserService.java
package com.rkostiuk.cstask.service;

import com.rkostiuk.cstask.dto.UserSearchRequest;
import com.rkostiuk.cstask.dto.UserAddressResponse;
import com.rkostiuk.cstask.repository.UserRepository;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.stereotype.Service;

@Service
public class DataJpaUserService implements UserService {
    private final UserRepository userRepository;

    public DataJpaUserService(UserRepository userRepository) {
        this.userRepository = userRepository;
    }

    @Override
    public Page<UserAddressResponse> findUsersWithBirthDateBetween(UserSearchRequest request, Pageable pageable) {
        return userRepository.findByBirthDateBetween(request.fromIncluding(), request.toExcluding(), pageable);
    }
}
